+++ app/src/main/java/com/dsource/idc/jellowintl/packageUpdate/UpdatePackageCheckUtils.java
@@ -1,11 +1,14 @@
 package com.dsource.idc.jellowintl.packageUpdate;
 
 import android.content.Context;
+import android.content.DialogInterface;
+import android.content.Intent;
+import android.widget.Button;
 
 import androidx.annotation.NonNull;
-import androidx.core.content.ContextCompat;
+import androidx.appcompat.app.AlertDialog;
 
-import com.dsource.idc.jellowintl.LanguageSelectActivity;
+import com.dsource.idc.jellowintl.LanguagePackUpdateActivity;
 import com.dsource.idc.jellowintl.R;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
@@ -50,25 +53,17 @@ public class UpdatePackageCheckUtils {
     }
 
     private void compareVersionCode(Context context) {
-        LanguageSelectActivity activity;
-        if(context instanceof LanguageSelectActivity) {
-            activity = (LanguageSelectActivity) context;
-        }else{
-            return;
-        }
         File fVersionCodeJSON = getVersionCodeMapJSON(context);
         File fOldVersionCodeJSON = getOldVersionCodeMapJSON(context);
         if (!fOldVersionCodeJSON.exists()){
-            activity.getScreenMenu().getItem(10).
-                setIcon(ContextCompat.getDrawable(activity, R.drawable.ic_launcher));
+            showUpdateDialog(context);
             return;
         }
 
         HashMap<String, String> newVersionCode = getHashMap(fVersionCodeJSON);
         HashMap<String, String> OldVersionCode = getHashMap(fOldVersionCodeJSON);
         if (!newVersionCode.get(KEY_VERSION_CODE).equals(OldVersionCode.get(KEY_VERSION_CODE))){
-            activity.getScreenMenu().getItem(10).
-                    setIcon(ContextCompat.getDrawable(activity, R.drawable.ic_launcher));
+            showUpdateDialog(context);
         }
     }
 
@@ -96,5 +91,42 @@ public class UpdatePackageCheckUtils {
         }
 
         return hashMap;
+    }
+
+
+    private void showUpdateDialog(final Context context){
+        String updateNow = context.getString(R.string.update_now);
+        String updateLater = context.getString(R.string.update_later);
+        String message = context.getString(R.string.update_message);
+
+        AlertDialog.Builder builder = new AlertDialog.Builder(context);
+        // Add the buttons
+        builder
+            .setPositiveButton(updateNow, new DialogInterface.OnClickListener() {
+                public void onClick(DialogInterface dialog, int id) {
+                    context.startActivity(new Intent(context, LanguagePackUpdateActivity.class));
+                    dialog.dismiss();
+                }
+            })
+            .setNegativeButton(updateLater, new DialogInterface.OnClickListener() {
+                @Override
+                public void onClick(DialogInterface dialog, int i) {
+                    dialog.dismiss();
+                }
+            })
+            // Set other dialog properties
+            .setCancelable(true)
+            .setMessage(message);
+
+        // Create the AlertDialog
+        AlertDialog dialog = builder.create();
+        // Show the AlertDialog
+        dialog.show();
+        Button positiveButton = dialog.getButton(AlertDialog.BUTTON_POSITIVE);
+        positiveButton.setTextColor(context.getResources().getColor(R.color.colorAccent));
+        positiveButton.setTextSize(18f);
+        Button negativeButton = dialog.getButton(AlertDialog.BUTTON_NEGATIVE);
+        negativeButton.setTextColor(context.getResources().getColor(R.color.colorAccent));
+        negativeButton.setTextSize(18f);
     }
 }